Added a cropper and a VideoMaker
The cropper takes in a video file and saves full frames from a given timestamp should you need it and also the ability reduce their size and save it in a 1:1 ratio

The VideoMaker joins the frames up into a video. On windows, DIVX and XVID work, both save video as an avi. Needs to be tested on Linux to get MKVs instead